requirements.txt
fastapi
uvicorn
python-dotenv
groq
tavily-python
pydantic
gunicorn